+++ src/Extensions/Exceptions/MixinInvalidException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Umbrellio\Postgres\Extensions\Exceptions;
+namespace Fuwasegu\Postgres\Extensions\Exceptions;
 
 use Exception;
