+++ src/Schema/Builders/Indexes/Unique/UniquePartialBuilder.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Umbrellio\Postgres\Schema\Builders\Indexes\Unique;
+namespace Fuwasegu\Postgres\Schema\Builders\Indexes\Unique;
 
 use Illuminate\Support\Fluent;
-use Umbrellio\Postgres\Schema\Builders\WhereBuilderTrait;
+use Fuwasegu\Postgres\Schema\Builders\WhereBuilderTrait;
 
 class UniquePartialBuilder extends Fluent
 {
